+++ tests/unit/ClientBatchTest.php
@@ -2,7 +2,6 @@
 
 namespace Domnikl\Test\Statsd;
 
-require_once __DIR__ . '/../lib/Client.php';
 require_once __DIR__ . '/ConnectionMock.php';
 
 use Domnikl\Statsd\Client as Client;