+++ tests/unit/Connection/UdpSocketTest.php
@@ -2,16 +2,13 @@
 
 namespace Domnikl\Test\Statsd\Connection;
 
-use Domnikl\Statsd\Connection\TcpSocket;
+use Domnikl\Statsd\Connection\UdpSocket;
 
-require_once __DIR__ . '/../../lib/Connection/InetSocket.php';
-require_once __DIR__ . '/../../lib/Connection/TcpSocket.php';
-
-class TcpSocketTest extends \PHPUnit_Framework_TestCase
+class UdpSocketTest extends \PHPUnit_Framework_TestCase
 {
     public function testInit()
     {
-        $connection = new TcpSocket('localhost', 8125, 10, true);
+        $connection = new UdpSocket('localhost', 8125, 10, true);
         $this->assertEquals('localhost', $connection->getHost());
         $this->assertEquals(8125, $connection->getPort());
         $this->assertEquals(10, $connection->getTimeout());
@@ -20,7 +17,7 @@ class TcpSocketTest extends \PHPUnit_Framework_TestCase
 
     public function testInitDefaults()
     {
-        $connection = new TcpSocket();
+        $connection = new UdpSocket();
         $this->assertEquals('localhost', $connection->getHost());
         $this->assertEquals(8125, $connection->getPort());
         $this->assertNull($connection->getTimeout());